refactor(provider): separate canonical subscription primitives
