+++ crates/chain-state/Cargo.toml
@@ -36,7 +36,6 @@ tracing.workspace = true
 
 # misc
 derive_more.workspace = true
-auto_impl.workspace = true
 metrics.workspace = true
 parking_lot.workspace = true
 pin-project.workspace = true
