+++ crates/rpc/rpc/src/reth.rs
@@ -17,8 +17,7 @@ use reth_primitives_traits::{NodePrimitives, SealedHeader};
 use reth_rpc_api::{RethApiServer, RethJitAction};
 use reth_rpc_eth_types::{EthApiError, EthResult};
 use reth_storage_api::{
-    BlockReader, BlockReaderIdExt, ChangeSetReader, NodePrimitivesProvider, StateProviderFactory,
-    TransactionVariant,
+    BlockReader, BlockReaderIdExt, ChangeSetReader, StateProviderFactory, TransactionVariant,
 };
 use reth_tasks::{pool::BlockingTaskGuard, Runtime};
 use serde::Serialize;
@@ -191,16 +190,16 @@ impl<Provider, EvmConfig> RethApiServer for RethApi<Provider, EvmConfig>
 where
     Provider: BlockReaderIdExt
         + ChangeSetReader
-        + NodePrimitivesProvider
         + StateProviderFactory
-        + BlockReader<Block = <<Provider as NodePrimitivesProvider>::Primitives as NodePrimitives>::Block>
-        + CanonStateSubscriptions
+        + BlockReader<
+            Block = <<Provider as CanonStateSubscriptions>::Primitives as NodePrimitives>::Block,
+        > + CanonStateSubscriptions
         + ForkChoiceSubscriptions<
-            Header = <<Provider as NodePrimitivesProvider>::Primitives as NodePrimitives>::BlockHeader,
+            Header = <<Provider as CanonStateSubscriptions>::Primitives as NodePrimitives>::BlockHeader,
         >
         + PersistedBlockSubscriptions
         + 'static,
-    EvmConfig: ConfigureEvm<Primitives = <Provider as NodePrimitivesProvider>::Primitives> + 'static,
+    EvmConfig: ConfigureEvm<Primitives = <Provider as CanonStateSubscriptions>::Primitives> + 'static,
 {
     /// Handler for `reth_getBalanceChangesInBlock`
     async fn reth_get_balance_changes_in_block(
